# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DevConnect REST API
  version: 1.0.0
  description: |
    Firebase-authenticated BFF over Firestore for the Developers Networking App.

    **Hybrid architecture:** Android uses the Firestore SDK for realtime features (chat, tasks, match).
    This REST layer provides analytics snapshots, admin tools, and Postman-friendly contracts.

    Preferred base path: `/api/v1`. Legacy `/api/*` aliases remain for the Android app
    (`GET /api/dashboard/stats`, `GET /api/projects`) without requiring frontend changes.
  contact:
    name: DevConnect Team
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://devconnect-api.onrender.com
    description: Production (optional — Render)
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
...
security:
  - FirebaseBearer: []
paths:
...
  /api/v1/conversations/{conversationId}/messages:
    get:
      tags: [Conversations]
      summary: Messages in conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
    post:
      tags: [Conversations]
      summary: Send message
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
      responses:
        '201':
          description: Message sent
...